feat: complete CI pipeline overhaul — all 9 improvements
CI/CD:
- Concurrency group cancels stale in-progress runs on same ref
- ruff --output-format github surfaces lint errors as inline PR annotations
- pytest --tb=short gives readable tracebacks without full call-stack spam
- pip-audit + pnpm audit steps scan for known CVEs on every PR
- New migrations job: spins up TimescaleDB 2.16.1-pg16, runs alembic
  upgrade head then alembic check — catches unapplied model changes
- Dependabot: weekly updates for pip, npm, and GitHub Actions

Frontend:
- vitest 2.1.9 with jsdom + @testing-library/react set up from scratch
- 61 tests across format.ts, nutrients.ts, portions.ts, SlopeChip

Backend:
- mypy added (clean, zero errors); fixed 6 real type errors found:
  - tasks.py: user_id was silently dropped from normalize_hae_payload call
  - today.py: tz str parameter shadowed by ZoneInfo reassignment
  - goals.py: cal_target int/float mismatch around max() call
  - hae_normalizer.py: iname str|None annotation conflict in loop scope
  - llm_client.py: untyped choices list
  - timescale.py: false-positive alembic.op stub gap

+++ backend/luma/api/goals.py
@@ -192,10 +192,10 @@ async def recommend_goals(user: CurrentUser, db: DbDep) -> dict[str, Any]:
 
     # Calorie target
     if target_weight_kg and target_weight_kg < current_weight_kg - 1.0:
-        cal_target = round((tdee - 500) / 50) * 50  # ~500 kcal/day deficit ≈ 0.45 kg (1 lb)/week
+        cal_target = float(round((tdee - 500) / 50) * 50)  # ~500 kcal/day deficit ≈ 0.45 kg (1 lb)/week
         mode = "deficit"
     else:
-        cal_target = round(tdee / 50) * 50
+        cal_target = float(round(tdee / 50) * 50)
         mode = "maintenance"
 
     # Mayo-style minimum — never recommend below the clinically advised floor.
